+++ include/shadefuncs.h
@@ -47,7 +47,7 @@
  *  The interface to the various material property & texture routines.
  */
 struct mfuncs {
-    unsigned long mf_magic;	/**< @brief To validate structure */
+    uint32_t mf_magic;		/**< @brief To validate structure */
     char *mf_name;		/**< @brief Keyword for material */
     struct mfuncs *mf_forw;	/**< @brief Forward link */
     int mf_inputs;		/**< @brief shadework inputs needed */
